Add repository interface for backgroundSync. Moved NotesBackgroundSync to sync package
