+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
 		<activity android:name=".create.CreateNoteActivity" />
 
 		<service
-			android:name=".notes.NotesBackgroundSync"
+			android:name=".sync.NotesBackgroundSync"
 			android:exported="true"
 			android:permission="com.google.android.gms.permission.BIND_NETWORK_TASK_SERVICE">
 			<intent-filter>
