+++ app/src/main/java/com/arconsis/mvvmnotesample/sync/NotesSyncRepository.kt
@@ -1,0 +1,11 @@
+package com.arconsis.mvvmnotesample.sync
+
+/**
+ * Created by akoufatzis on 18.05.17.
+ */
+interface NotesSyncRepository {
+
+    fun schedule()
+    fun unschedule()
+    fun notify(notificationHandler: (() -> Unit)?)
+}
